api/requirements.txt
openai==1.3.0
google-generativeai==0.8.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
